Create REST_API_EXAMPLE.mq5
This is a template I wrote of how to connect to an API using MQL5. MQL5 is does have supporting documentation but a lot of it is outdated and I managed to get it working after spending a few hours on it.

I used a public API to test the code.